requirements.txt
amqp==5.3.1
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
CacheControl==0.14.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==46.0.5
Django==6.0.3
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
firebase_admin==7.2.0
google-api-core==2.30.0
google-auth==2.49.0
google-cloud-core==2.5.0
google-cloud-firestore==2.24.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
msgpack==1.1.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
